RAC-304 fix : 자동 갱신 클래스 분리
